% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{abs_data_up_to_date}
\alias{abs_data_up_to_date}
\title{Check if an ABS data source is up to date}
\usage{
abs_data_up_to_date(cat_no, data_name = NULL)
}
\arguments{
\item{cat_no}{string. include the ".0"}

\item{data_name}{optional}
}
\value{
logical
}
\description{
Check if an ABS data source is up to date
}
