% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{abs_next_release}
\alias{abs_next_release}
\title{Find the date of the next release of an ABS dataset}
\usage{
abs_next_release(cat_no)
}
\arguments{
\item{cat_no}{}
}
\value{

}
\description{
Find the date of the next release of an ABS dataset
}
