% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abs_urls}
\alias{abs_urls}
\title{Match ABS Catalogue names with URLS}
\format{
An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
abs_urls
}
\description{
Match ABS Catalogue names with URLS
}
\keyword{datasets}
