% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aitidata_catalogues}
\alias{aitidata_catalogues}
\title{aitidata Catalogues}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 7 columns.
}
\usage{
aitidata_catalogues
}
\description{
aitidata Catalogues
}
\keyword{datasets}
