% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzsic}
\alias{anzsic}
\title{Australian and New Zealand Standard Industrial Classification (ANZSIC)}
\format{
A dataframe with 506 rows and 4 variables:
\describe{
\item{division}{industry division}
\item{subdivision}{subdivision of division}
\item{group}{group of subdivision}
\item{class}{class of group}
}
}
\source{
\url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/1292.0.55.002Main+Features12006?OpenDocument}
}
\usage{
anzsic
}
\description{
A dataset containing the industry divisions, subdivisions, groups, and classes of the
2006 ANZSIC organised heirarchically
}
\keyword{datasets}
