% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_manufacturing}
\alias{aus_manufacturing}
\title{Australian Manufacturing}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2604 rows and 5 columns.
}
\usage{
aus_manufacturing
}
\description{
Australian Manufacturing
}
\keyword{datasets}
