% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabee_sa2}
\alias{cabee_sa2}
\title{Counts of Australian Businesses, including Entries and Exists, June 2015 to June 2019}
\format{
A dataframe with 6 variables:
\describe{
\item{date}{date}
\item{division}{Industry division of operating business}
\item{sa2_main_2016}{9 digit statistical area code (2016 SA2)}
\item{sa2_name_2016}{Name of statistical area}
\item{indicator}{Employment ranges}
}
}
\source{
\url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/8165.0Main+Features2June\%202015\%20to\%20June\%202019?OpenDocument}
}
\usage{
cabee_sa2
}
\description{
A dataset containing the number of businesses, by industry division, by SA2 statistical areas,
by employment size ranges
}
\keyword{datasets}
