% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{check_data_up_to_date}
\alias{check_data_up_to_date}
\title{Checks that installed version of daitir has
up to date data}
\usage{
check_data_up_to_date()
}
\value{
A message describing the up-to-date-ness of the installed package.
}
\description{
Checks that installed version of daitir has
up to date data
}
