% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{COVID-19 Geographic data}
\format{
A dataframe with 6 variables:
\describe{
\item{date}{Date}
\item{state_name_2016}{State name}
\item{indicator}{COVID-19 indicators}
\item{value}{Value}
\item{statistical_area}{Number of digits in statistical area}
\item{statistical_area_code}{Statistical area code}
}
}
\usage{
covid_data
}
\description{
A dataset containing the jobkeeper applications, jobseeker payments, and derived impact from COVID-19 at
a geographical level. Jobkeeper and Jobseeker data is based on the SA2 classiciation, and employment
impact is based on the SA4 classification. This dataset combines data from jobseeker_sa2, jobkeeper_sa2,
payroll_sa4, and small_area_labour_market datasets and is intended for use on the AITI Economic
Indicators dashboard.
}
\keyword{datasets}
