% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{current_release}
\alias{current_release}
\title{a URL, or catalogue number from the ABS.
Can also look up release dates of data on data.gov.au via the \code{source} parameter.}
\usage{
current_release(cat_string = NULL, url = NULL, source = "abs")
}
\arguments{
\item{cat_string}{string (optional)}

\item{url}{string (optional)}

\item{source}{string}
}
\value{
a date
}
\description{
a URL, or catalogue number from the ABS.
Can also look up release dates of data on data.gov.au via the \code{source} parameter.
}
\examples{
\dontrun{current_release(cat_string = "labour-force-australia")}
}
