% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_cube.R
\name{download_data_cube}
\alias{download_data_cube}
\title{The filepath is returned invisibly which enables piping to \code{unzip()} or \code{readxl::read_excel}.}
\usage{
download_data_cube(catalogue_string, cube, path = here::here("data-raw"))
}
\arguments{
\item{catalogue_string}{string}

\item{cube}{string}

\item{path}{string}
}
\description{
The filepath is returned invisibly which enables piping to \code{unzip()} or \code{readxl::read_excel}.
}
