% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download a file from the internet}
\usage{
download_file(file_download_url, path = here::here("data-raw"))
}
\arguments{
\item{file_download_url}{url of file to download}

\item{path}{path to download the file to}
}
\value{
Returns (silently) the path where the file has been downloaded
}
\description{
Download a file from the internet
}
