% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_complexity}
\alias{economic_complexity}
\title{Australian Subnational Economic Complexity}
\format{
A dataframe with 704 observations and 4 variables
\describe{
\item{location_code}{state or territory abbreviation}
\item{year}{Year}
\item{indicator}{Place specific economic complexity indicator (diversity, eci, coi, total exports)}
\item{value}{Value of indicator}
}
}
\usage{
economic_complexity
}
\description{
A dataset containing economic complexity indicators
for Australian states and territories
}
\keyword{datasets}
