% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employment_by_industry_detailed}
\alias{employment_by_industry_detailed}
\title{ABS Detailed employment by industry}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 727764 rows and 7 columns.
}
\usage{
employment_by_industry_detailed
}
\description{
ABS Detailed employment by industry
}
\keyword{datasets}
