% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{gov_current_release}
\alias{gov_current_release}
\title{Current release date from data.gov}
\usage{
gov_current_release(url = url)
}
\arguments{
\item{url}{string}
}
\value{
Date
}
\description{
Current release date from data.gov
}
