% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_spending}
\alias{household_spending}
\title{Experimental household spending}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6426 rows and 9 columns.
}
\usage{
household_spending
}
\description{
Experimental household spending
}
\keyword{datasets}
