% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industry_employment_detailed}
\alias{industry_employment_detailed}
\title{ABS Detailed employment by industry}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 804224 rows and 7 columns.
}
\usage{
industry_employment_detailed
}
\description{
ABS Detailed employment by industry
}
\keyword{datasets}
