% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industry_value_add}
\alias{industry_value_add}
\title{Industry value add}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51051 rows and 7 columns.
}
\usage{
industry_value_add
}
\description{
Industry value add
}
\keyword{datasets}
