% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_vacancies_index}
\alias{internet_vacancies_index}
\title{ABS Employment by Industry}
\format{
A dataframe with 11 variables:
\describe{
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{indicator}{Indicator}
\itemize{
\item{Employed total}
\item{Underemployed total}
\item{Underemployment ratio (proportion of employed)}
\item{Employed full-time}
\item{Employed part-time}
}
\item{industry}{Industry of employment}
\item{gender}{Gender}
\item{age}{Age}
\item{state}{State}
\item{series_type}{Level of adjustment of original data: original (no adjustment), seasonally adjusted, or trend}
\item{value}{Value of employment indicator}
\item{unit}{Unit of measurement for indiciator}
}

A dataframe with 7 variables
\describe{
\item{state}{State}
\item{date}{year month date, since January 2006}
\item{anzsco_2}{2 digit ANZSCO}
\item{occupation}{Occupation name}
\item{value}{Number of internet vacancies recorded}
\item{anzsco_1}{1 digit ANZSCO}
\item{occupation_group}{Occupation group name}
}
}
\source{
\url{https://www.abs.gov.au/ausstats/abs@.nsf/PrimaryMainFeatures/6291.0.55.001?OpenDocument}
ABS Detailed employment by industry
Internet Vacancies Index

A dataset containing monthly nation and sub-national internet vacancies by
occupation since January 2006. Data is updated monthly, about one week
after the Labour Force Survey is released.

\url{https://lmip.gov.au/default.aspx?LMIP/GainInsights/VacancyReport}
}
\usage{
internet_vacancies_index
}
\description{
A dataset containing quarterly national and sub-national employment indicators
by industry of employment since February 1991.
}
\keyword{datasets}
