% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_vacancies_regional}
\alias{internet_vacancies_regional}
\title{Internet Vacancies Regional}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 288933 rows and 7 columns.
}
\usage{
internet_vacancies_regional
}
\description{
Internet Vacancies Regional
}
\keyword{datasets}
