% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_vacancy_index}
\alias{internet_vacancy_index}
\title{Internet Vacancies Index}
\format{
A dataframe with 7 variables
\describe{
\item{state}{State}
\item{date}{year month date, since January 2006}
\item{anzsco_2}{2 digit ANZSCO}
\item{occupation}{Occupation name}
\item{value}{Number of internet vacancies recorded}
\item{anzsco_1}{1 digit ANZSCO}
\item{occupation_group}{Occupation group name}
}
}
\source{
\url{https://lmip.gov.au/default.aspx?LMIP/GainInsights/VacancyReport}
}
\usage{
internet_vacancy_index
}
\description{
A dataset containing monthly nation and sub-national internet vacancies by
occupation since January 2006. Data is updated monthly, about one week
after the Labour Force Survey is released.
}
\keyword{datasets}
