% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_vacancy_regions}
\alias{internet_vacancy_regions}
\title{Internet vacancy regions}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 37 rows and 2 columns.
}
\usage{
internet_vacancy_regions
}
\description{
Internet vacancy regions
}
\keyword{datasets}
