% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobseeker_sa2}
\alias{jobseeker_sa2}
\title{JobSeeker data by SA2}
\format{
A data frame with 4 variables
\describe{
\item{sa2_code}{Statistical Area 2 Code. A mix of 2016 and 2021 boundaries}
\item{sa2_name}{Statistical Area 2 Name. A mix of 2016 and 2021 boundaries}
\item{date}{Date}
\item{indicator}{Jobseeker payment, Youth allowance other, Jobseeker growth, Youth allowance growth}
\item{value}{Value}
}
}
\source{
\url{https://data.gov.au/data/dataset/jobseeker-payment-and-youth-allowance-recipients-monthly-profile}
}
\usage{
jobseeker_sa2
}
\description{
Department of Social Services estimates of monthly jobseeker and
youth allowance payments since March 2020
}
\keyword{datasets}
