% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labour_account}
\alias{labour_account}
\title{ABS Labour Account}
\format{
A dataframe with 10 variables
\describe{
\item{date}{Date}
\item{month}{Month}
\item{year}{Year}
\item{prefix}{Prefix on the indicator - safe to ignore}
\item{indicator}{Labour account indicator}
\item{industry}{Industry}
\item{series_type}{Series type}
\item{value}{Value}
\item{unit}{Unit of value}
}
}
\source{
\url{https://www.abs.gov.au/ausstats/abs@.nsf/mf/6150.0.55.003}
}
\usage{
labour_account
}
\description{
Experimental ABS labour force data
}
\keyword{datasets}
