% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobility_google}
\alias{mobility_google}
\title{Google mobility}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35532 rows and 4 columns.
}
\usage{
mobility_google
}
\description{
Google mobility
}
\keyword{datasets}
