% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_underemployment}
\alias{occupation_underemployment}
\title{Underemployment by occupation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6144 rows and 11 columns.
}
\usage{
occupation_underemployment
}
\description{
Underemployment by occupation
}
\keyword{datasets}
