% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{payroll_substate}
\alias{payroll_substate}
\title{Payroll Substate}
\format{
A dataframe with 5 variables:
\describe{
\item{state_name_2016}{State}
\item{date}{Date}
\item{value}{Number of payroll jobs - indexed to March 14th 2020}
\item{indicator}{This dataset contains only the payroll index data}
}
}
\source{
\url{https://www.abs.gov.au/ausstats/abs@.nsf/mf/6160.0.55.001}
}
\usage{
payroll_substate
}
\description{
ABS Weekly Payroll data for Statistical Area 4 Regions
}
\keyword{datasets}
