% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retail_trade}
\alias{retail_trade}
\title{Retail trade}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 86258 rows and 9 columns.
}
\usage{
retail_trade
}
\description{
Retail trade
}
\keyword{datasets}
