% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seek_data}
\alias{seek_data}
\title{Seek Job Ads}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8500 rows and 8 columns.
}
\usage{
seek_data
}
\description{
Seek Job Ads
}
\keyword{datasets}
