% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_area_labour_market}
\alias{small_area_labour_market}
\title{Small Area Labour Market}
\format{
A dataframe with 6 variables:
\describe{
\item{indicator}{One of Labour force total, Unemployment rate, Unemployed total}
\item{sa2_name_2016}{Name of the SA2 region}
\item{sa2_main_2016}{9 digit code representing the SA2 region}
\item{date}{Quarterly since December 2010}
\item{value}{Value of the indicator}
\item{state_name_2016}{State name}
}
}
\source{
\url{https://www.employment.gov.au/small-area-labour-markets-publication-0}
}
\usage{
small_area_labour_market
}
\description{
The Department of Education, Skills and Employment quarterly regional
estimates of unemployment and the unemployment rate at the Statistical
Area Level 2 (SA2)
}
\keyword{datasets}
