% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{south_australia_net_debt}
\alias{south_australia_net_debt}
\title{SA Net Debt}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 5 columns.
}
\usage{
south_australia_net_debt
}
\description{
SA Net Debt
}
\keyword{datasets}
