% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{underemployment_by_industry}
\alias{underemployment_by_industry}
\title{Underemployment by industry}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14136 rows and 11 columns.
}
\usage{
underemployment_by_industry
}
\description{
Underemployment by industry
}
\keyword{datasets}
