% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{underemployment_by_occupation}
\alias{underemployment_by_occupation}
\title{Underemployment by occupation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5952 rows and 11 columns.
}
\usage{
underemployment_by_occupation
}
\description{
Underemployment by occupation
}
\keyword{datasets}
