% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{underutilisation}
\alias{underutilisation}
\title{ABS Measures of Underutilisation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15912 rows and 10 columns.
}
\usage{
underutilisation
}
\description{
ABS Measures of Underutilisation
}
\keyword{datasets}
