% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aus_manufacturing.R
\name{update_aus_manufacturing}
\alias{update_aus_manufacturing}
\title{Update Australian Manufacturing Data}
\usage{
update_aus_manufacturing(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update Australian Manufacturing Data
}
