% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cabee_sa2.R
\name{update_cabee_sa2}
\alias{update_cabee_sa2}
\title{Update Count of Australian Businesses Data by SA2}
\usage{
update_cabee_sa2(force_update = FALSE)
}
\arguments{
\item{force_update}{logical. FALSE (the default) checks if new data is available before updating.}
}
\value{
logical. TRUE if the update was successful
}
\description{
Update Count of Australian Businesses Data by SA2
}
