% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_by_industry.R
\name{update_employment_by_industry}
\alias{update_employment_by_industry}
\title{Update Employment by industry data}
\usage{
update_employment_by_industry(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update Employment by industry data
}
