% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_household_spending.R
\name{update_household_spending}
\alias{update_household_spending}
\title{Update Household spending data}
\usage{
update_household_spending(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical. TRUE if the update was successful
}
\description{
Update Household spending data
}
