% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_industry_employment.R
\name{update_industry_employment}
\alias{update_industry_employment}
\title{Update Employment by industry data}
\usage{
update_industry_employment(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update Employment by industry data
}
