% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_industry_employment_detailed.R
\name{update_industry_employment_detailed}
\alias{update_industry_employment_detailed}
\title{Update detailed employment by industry data}
\usage{
update_industry_employment_detailed(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if data updated successfully
}
\description{
Update detailed employment by industry data
}
