% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_industry_underemployment.R
\name{update_industry_underemployment}
\alias{update_industry_underemployment}
\title{Update \code{underemployment_by_industry} dataset.}
\usage{
update_industry_underemployment(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if update was successful
}
\description{
Update \code{underemployment_by_industry} dataset.
}
