% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_internet_vacancies_index.R
\name{update_internet_vacancies_index}
\alias{update_internet_vacancies_index}
\title{Update internet vacancies index}
\usage{
update_internet_vacancies_index(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if the data updated successfully
}
\description{
Update internet vacancies index
}
