% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_internet_vacancies_regional.R
\name{update_internet_vacancies_regional}
\alias{update_internet_vacancies_regional}
\title{prepare \code{internet_vacancies_regional} dataset}
\usage{
update_internet_vacancies_regional(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if the update succeeded
}
\description{
prepare \code{internet_vacancies_regional} dataset
}
