% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_jobseeker_sa2.R
\name{update_jobseeker_sa2}
\alias{update_jobseeker_sa2}
\title{Update JobSeeker SA2 Data if available}
\usage{
update_jobseeker_sa2(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update JobSeeker SA2 Data if available
}
