% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_labour_force.R
\name{update_labour_force}
\alias{update_labour_force}
\title{Update labour force data}
\usage{
update_labour_force(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if data updated successfully
}
\description{
Update labour force data
}
