% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_national_accounts.R
\name{update_national_accounts}
\alias{update_national_accounts}
\title{Update National Accounts ABS Data}
\usage{
update_national_accounts(force_update = FALSE)
}
\arguments{
\item{force_update}{logical. FALSE (the default) checks if new data is available before updating.}
}
\value{
logical. TRUE if the update completed successfully or no new data found
}
\description{
Update National Accounts ABS Data
}
