% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retail_trade.R
\name{update_retail_trade}
\alias{update_retail_trade}
\title{Update Retail Trade dataset}
\usage{
update_retail_trade(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update Retail Trade dataset
}
