% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_area_labour_market.R
\name{update_small_area_labour_market}
\alias{update_small_area_labour_market}
\title{Update small area labour market data}
\usage{
update_small_area_labour_market(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical
}
\description{
Update small area labour market data
}
