% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underemployment_by_industry.R
\name{update_underemployment_by_industry}
\alias{update_underemployment_by_industry}
\title{Update \code{underemployment_by_industry} dataset.}
\usage{
update_underemployment_by_industry(force_update = FALSE)
}
\arguments{
\item{force_update}{logical}
}
\value{
logical: TRUE if update was successful
}
\description{
Update \code{underemployment_by_industry} dataset.
}
