% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{abs_current_release}
\alias{abs_current_release}
\title{Find the release date of a dataset from the ABS.}
\usage{
abs_current_release(cat_string = NULL, url = NULL)
}
\arguments{
\item{cat_string}{string}

\item{url}{string}
}
\description{
Find the release date of a dataset from the ABS.
}
\examples{
\dontrun{current_release(cat_string = "labour-force-australia")}
}
