% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{abs_current_release}
\alias{abs_current_release}
\title{Find the release date of the most recent ABS Catalogue}
\usage{
abs_current_release(cat_no)
}
\arguments{
\item{cat_no}{}
}
\value{
date
}
\description{
Find the release date of the most recent ABS Catalogue
}
