% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{abs_next_release}
\alias{abs_next_release}
\title{Find the date of the next release of an ABS dataset}
\usage{
abs_next_release(cat_string = NULL, url = NULL)
}
\arguments{
\item{cat_string}{string}

\item{url}{string}
}
\value{
date
}
\description{
Find the date of the next release of an ABS dataset
}
