% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download a file from the internet}
\usage{
download_file(file_download_url, path = here::here("data-raw"))
}
\arguments{
\item{path}{}

\item{download_url}{}
}
\value{
Returns the path where the file has been downloaded
}
\description{
Download a file from the internet
}
