% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobkeeper_state}
\alias{jobkeeper_state}
\title{Jobkeeper applications by state}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 297 rows and 10 columns.
}
\usage{
jobkeeper_state
}
\description{
Jobkeeper applications by state
}
\keyword{datasets}
