% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobkeeper_state}
\alias{jobkeeper_state}
\title{Jobkeeper}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 216 rows and 8 columns.
}
\usage{
jobkeeper_state
}
\description{
Jobkeeper
}
\keyword{datasets}
