% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobseeker_state}
\alias{jobseeker_state}
\title{jobseeker state}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 414 rows and 10 columns.
}
\usage{
jobseeker_state
}
\description{
jobseeker state
}
\keyword{datasets}
