% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobseeker_state}
\alias{jobseeker_state}
\title{Jobseeker}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 234 rows and 8 columns.
}
\usage{
jobseeker_state
}
\description{
Jobseeker
}
\keyword{datasets}
