% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobility_facebook}
\alias{mobility_facebook}
\title{Facebook Mobility}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 194454 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 194454 rows and 6 columns.
}
\usage{
mobility_facebook

mobility_facebook
}
\description{
Facebook Mobility

Facebook mobility
}
\keyword{datasets}
