% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{payroll_index}
\alias{payroll_index}
\title{ABS Weekly Payroll Index}
\format{
A dataframe with 6 variables:
\describe{
\item{date}{Date}
\item{gender}{Gender}
\item{age}{Age group}
\item{state}{State}
\item{industry}{Industry}
\item{value}{Value}
}
}
\source{
\url{https://www.abs.gov.au/ausstats/abs@.nsf/mf/6160.0.55.001}
}
\usage{
payroll_index
}
\description{
ABS Weekly Payroll data since 4th January 2020.
}
\keyword{datasets}
